Parallelization of Boruvka's Minimum Spanning Tree Algorithm using ReentrantLock
Boruvka's algorithm computes the minimal spanning tree through successive applications of edge-contraction on an input graph (without self-loops).  In edge-contraction, an edge is chosen from the graph and a new node is formed with the union of the connectivity of the incident nodes of the chosen edge.  In the case that there are duplicate edges,  only the one with least weight is carried through in the union.  Figure 2 demonstrates this process.  Boruvka's algorithm proceeds in an unordered fashion.  Each node performs edge contraction with its lightest neighbor.

### Work Done:

- Implemented the parallel and sequential version of Boruvka's Minimum Spanning Tree Algorithm, and wrote test cases to calculate speedup achieved in parallel version.

## Testing:

- Unit tested in Apple M2 Pro 16 GB (Test can fail based on number of threads, change the expectedSpeedup accordingly)